share view code between home page and section page
