+++ samples/ControlCatalog/CustomThemes.xaml
@@ -7,14 +7,52 @@
       <Color x:Key="HamburgerChromeMediumColor">#FF1F1F1F</Color>
       <Color x:Key="HamburgerAltHighColor">#FF000000</Color>
       <Color x:Key="HamburgerChromeLowColor">#FF171717</Color>
+      <SolidColorBrush x:Key="HomeCardBackground" Color="#2B2B33" />
+      <SolidColorBrush x:Key="HomeCardBorderBrush" Color="#3A3A45" />
+      <SolidColorBrush x:Key="HomeCardBorderBrushHover" Color="#8AB4FF" />
+      <SolidColorBrush x:Key="HomeSubtleForeground" Color="#9C9CAB" />
+      <SolidColorBrush x:Key="HomeBadgeBackground" Color="#31395A" />
+      <SolidColorBrush x:Key="HomeBadgeForeground" Color="#8AB4FF" />
     </ResourceDictionary>
     <ResourceDictionary x:Key="Default">
       <Color x:Key="HamburgerBaseHighColor">#99000000</Color>
       <Color x:Key="HamburgerChromeMediumColor">#FFE6E6E6</Color>
       <Color x:Key="HamburgerAltHighColor">#FFFFFFFF</Color>
       <Color x:Key="HamburgerChromeLowColor">#FFF2F2F2</Color>
+      <SolidColorBrush x:Key="HomeCardBackground" Color="#FFFFFF" />
+      <SolidColorBrush x:Key="HomeCardBorderBrush" Color="#E5E5EA" />
+      <SolidColorBrush x:Key="HomeCardBorderBrushHover" Color="#3E6DF3" />
+      <SolidColorBrush x:Key="HomeSubtleForeground" Color="#64646E" />
+      <SolidColorBrush x:Key="HomeBadgeBackground" Color="#E8EFFD" />
+      <SolidColorBrush x:Key="HomeBadgeForeground" Color="#3462D9" />
     </ResourceDictionary>
   </ResourceDictionary.ThemeDictionaries>
+
+  <ControlTheme x:Key="ButtonCardTheme" TargetType="Button">
+    <Setter Property="HorizontalAlignment" Value="Stretch" />
+    <Setter Property="VerticalAlignment" Value="Stretch" />
+    <Setter Property="Background" Value="{DynamicResource HomeCardBackground}" />
+    <Setter Property="BorderBrush" Value="{DynamicResource HomeCardBorderBrush}" />
+    <Setter Property="BorderThickness" Value="1" />
+    <Setter Property="CornerRadius" Value="10" />
+    <Setter Property="Padding" Value="16, 14" />
+    <Setter Property="MinHeight" Value="118" />
+    <Setter Property="Cursor" Value="Hand" />
+
+    <Style Selector="^ /template/ ContentPresenter">
+      <Setter Property="BoxShadow" Value="0 1 3 0 #0D000000" />
+      <Setter Property="Transitions">
+        <Transitions>
+          <BrushTransition Property="BorderBrush" Duration="0:0:0.1" />
+        </Transitions>
+      </Setter>
+    </Style>
+
+    <Style Selector="^:pointerover /template/ ContentPresenter">
+      <Setter Property="BorderBrush" Value="{DynamicResource HomeCardBorderBrushHover}" />
+    </Style>
+
+  </ControlTheme>
 
   <x:Double x:Key="NavigationItemHeight">36</x:Double>
   <BoxShadows x:Key="NavigationItemShadow">1 1 1 1 #2000, 0 0 1 1 #2fff</BoxShadows>
