+++ samples/ControlCatalog/Pages/SectionPage.axaml.cs
@@ -19,16 +19,5 @@ namespace ControlCatalog.Pages
         {
             InitializeComponent();
         }
-
-        private void UniformGrid_OnSizeChanged(object? sender, SizeChangedEventArgs e)
-        {
-            if (sender is not UniformGrid grid)
-                return;
-
-            const int minItemWidth = 248;
-            grid.Columns = Math.Max(
-                1,
-                (int)((e.NewSize.Width + grid.ColumnSpacing) / (minItemWidth + grid.ColumnSpacing)));
-        }
     }
 }
